+++ misago/conf/admin/forms/__init__.py
@@ -1,0 +1,6 @@
+from .analytics import ChangeAnalyticsSettingsForm
+from .base import ChangeSettingsForm
+from .captcha import ChangeCaptchaSettingsForm
+from .general import ChangeGeneralSettingsForm
+from .threads import ChangeThreadsSettingsForm
+from .users import ChangeUsersSettingsForm
