+++ misago/templates/misago/admin/form/input.html
@@ -6,7 +6,7 @@
 {% elif field|is_multiple_choice_field %}
   {% include "misago/admin/form/multiple_choice.html" %}
 {% elif field|is_textarea_field %}
-  <textarea {% render_attrs widget class_name="form-control" %} id="{{ field.id_for_label }}" name="{{ widget.name }}" {% render_bool_attrs widget %}>{% if widget.value is not None %}{{ widget.value }}{% endif %}</textarea>
+  <textarea {% render_attrs widget class_name="form-control" %} id="{{ field.id_for_label }}" name="{{ widget.name }}" {% render_bool_attrs widget %} {% if field.field.disabled %}disabled{% endif %}>{% if widget.value is not None %}{{ widget.value }}{% endif %}</textarea>
 {% else %}
-  <input {% render_attrs widget class_name="form-control" %} id="{{ field.id_for_label }}" name="{{ widget.name }}" type="{{ widget.type }}"{% if widget.value is not None %} value="{{ widget.value }}"{% endif %} {% render_bool_attrs widget %}>
+  <input {% render_attrs widget class_name="form-control" %} id="{{ field.id_for_label }}" name="{{ widget.name }}" type="{{ widget.type }}"{% if widget.value is not None %} value="{{ widget.value }}"{% endif %} {% render_bool_attrs widget %} {% if field.field.disabled %}disabled{% endif %}>
 {% endif %}
